Fix precision of value in skimage shape contours
When using the `skimage` method for producing shapefile contours, the `value` attribute attached to these contours was being rounded to the nearest integer, despite the precision being set to `float:2.1`. Clearly this means I don't understand exactly what fiona's precision strings mean; but I've modified it to match the `pcontour` schema and this fixes the unwanted rounding.